+++ const.go
@@ -1,8 +1,18 @@
-// Copyright IBM Corp. 2018, 2025
+// Copyright IBM Corp. 2018, 2026
 // SPDX-License-Identifier: MPL-2.0
 
 package tfe
 
 const (
+	// AuthenticationTokensPath is the API path for authentication tokens.
 	AuthenticationTokensPath = "authentication-tokens/%s"
+
+	// AdminSCIMTokensPath is the API path for admin SCIM tokens.
+	AdminSCIMTokensPath = "admin/scim-tokens"
+
+	// AdminSCIMGroupsPath is the API path for admin SCIM groups.
+	AdminSCIMGroupsPath = "admin/scim-groups"
+
+	// AdminSCIMGroupMappingPath is the API path for admin SCIM group mapping.
+	AdminSCIMGroupMappingPath = "admin/teams/%s/scim-group-mapping"
 )